Feat: 댓글 데이터 조회 fetchCommentData api 함수 추가 및 각 api 함수 게시물, 댓글이 존재하는지 확인 하는 로직과 예외처리하는 로직 추가 답글 모달창이 닫힐 때 해당 댓글의 답글 카운터를 최신 상태로 업데이트 시켜주기 위해서 추가 게시물이나 댓글이 삭제 됬을경우 발생하는 에러처리가 필요하여 커스텀 에러를 만들어서 예외처리

+++ src/api/firebase/commentAPI.ts
@@ -5,6 +5,7 @@ import {
   collection,
   deleteDoc,
   doc,
+  getDoc,
   getDocs,
   increment,
   limit,
@@ -22,6 +23,22 @@ import { getAuth } from "firebase/auth";
 const auth = getAuth();
 
 /**
+ * 댓글 데이터 조회
+ */
+export const fetchCommentData = async (
+  commentId: string
+): Promise<ICommentData | undefined> => {
+  try {
+    const commentsDoc = doc(db, `comments/${commentId}`);
+    const res = await getDoc(commentsDoc);
+    const data = res.data();
+    return data as ICommentData;
+  } catch (error) {
+    console.error(error);
+    throw error;
+  }
+};
+/**
  * 댓글 첫 페이지 조회
  */
 export const fetchFirstPageCommentData = async (
@@ -29,6 +46,13 @@ export const fetchFirstPageCommentData = async (
   pagePerData: number
 ) => {
   try {
+    // 게시물 확인 및 예외처리
+    const postDoc = doc(db, `post/${postId}`);
+    const postDocSnapshot = await getDoc(postDoc);
+    if(!postDocSnapshot.exists()){
+      throw new Error("게시물이 존재하지 않습니다.");
+    }
+
     const commentsRef = collection(db, "comments");
     const q = query(
       commentsRef,
@@ -117,6 +141,15 @@ export const fetchPagingCommentData = async (
  */
 export const fetchAddComment = async (commentData: ICommentData) => {
   try {
+    // 게시물 확인 작업 추가
+    const postDoc = doc(db, `post/${commentData.postId}`);
+    const postDocSnapshot = await getDoc(postDoc);
+
+    // 결과 확인 및 예외 처리
+    if (!postDocSnapshot.exists()) {
+      throw new Error("게시물이 존재하지 않습니다.");
+    }
+
     // comment colledcion에 댓글 doc 추가
     const commentsRef = collection(db, "comments");
     const addCommentPromise = setDoc(
@@ -125,7 +158,6 @@ export const fetchAddComment = async (commentData: ICommentData) => {
     );
 
     // 게시물 doc에 commenCount 늘리기
-    const postDoc = doc(db, `post/${commentData.postId}`);
     const addCommentCountPromise = updateDoc(postDoc, {
       commentCount: increment(1)
     });
@@ -141,10 +173,32 @@ export const fetchAddComment = async (commentData: ICommentData) => {
  * 댓글 수정
  */
 export const fetchEditComment = async (
-  commentEditData: Pick<ICommentData, "commentId" | "content">
+  commentEditData: Pick<ICommentData, "commentId" | "content" | "postId">
 ) => {
   try {
+    // 병렬 처리할 비동기 작업 배열 생성
+    const tasks = [];
+
+    // 게시물 확인 작업 추가
+    const postDoc = doc(db, `post/${commentEditData.postId}`);
+    tasks.push(getDoc(postDoc));
+
+    // 댓글 확인 작업 추가
     const commentDoc = doc(db, `comments/${commentEditData.commentId}`);
+    tasks.push(getDoc(commentDoc));
+
+    // 모든 작업을 병렬로 실행하고 결과 배열을 얻습니다.
+    const results = await Promise.all(tasks);
+
+    // 결과 확인 및 예외 처리
+    if (!results[0].exists()) {
+      throw new Error("게시물이 존재하지 않습니다.");
+    }
+
+    if (!results[1].exists()) {
+      throw new Error("댓글이 존재하지 않습니다.");
+    }
+
     await updateDoc(commentDoc, { content: commentEditData.content });
   } catch (error) {
     console.error(error);
@@ -159,12 +213,34 @@ export const fetchRemoveComment = async (
   commentData: ICommentData
 ): Promise<void> => {
   try {
+    // 병렬 처리할 비동기 작업 배열 생성
+    const tasks = [];
+
+    // 게시물 확인 작업 추가
+    const postDoc = doc(db, `post/${commentData.postId}`);
+    tasks.push(getDoc(postDoc));
+
+    // 댓글 확인 작업 추가
+    const commentDoc = doc(db, `comments/${commentData.commentId}`);
+    tasks.push(getDoc(commentDoc));
+
+    // 모든 작업을 병렬로 실행하고 결과 배열을 얻습니다.
+    const results = await Promise.all(tasks);
+
+    // 결과 확인 및 예외 처리
+    if (!results[0].exists()) {
+      throw new Error("게시물이 존재하지 않습니다.");
+    }
+
+    if (!results[1].exists()) {
+      throw new Error("댓글이 존재하지 않습니다.");
+    }
+
     // 댓글 삭제
     const commentsDoc = doc(db, `comments/${commentData.commentId}`);
     const removeCommentPromise = deleteDoc(commentsDoc);
 
     // 게시물 doc에서 commentCount 줄이기
-    const postDoc = doc(db, `post/${commentData.postId}`);
     const removeCommentCountPromise = updateDoc(postDoc, {
       commentCount: increment(-1)
     });
@@ -178,17 +254,41 @@ export const fetchRemoveComment = async (
 /**
  * 댓글 신고
  */
-
 export const fetchReportComment = async (
   reportCommentData: Pick<ICommentData, "commentId" | "reportCount"> & {
     postId: string;
   }
 ) => {
   try {
+    if (!auth.currentUser) return;
+
+    // 병렬 처리할 비동기 작업 배열 생성
+    const tasks = [];
+
+    // 게시물 확인 작업 추가
+    const postDoc = doc(db, `post/${reportCommentData.postId}`);
+    tasks.push(getDoc(postDoc));
+
+    // 댓글 확인 작업 추가
     const commentDoc = doc(db, `comments/${reportCommentData.commentId}`);
+    tasks.push(getDoc(commentDoc));
+
+    // 모든 작업을 병렬로 실행하고 결과 배열을 얻습니다.
+    const results = await Promise.all(tasks);
+
+    // 결과 확인 및 예외 처리
+    if (!results[0].exists()) {
+      throw new Error("게시물이 존재하지 않습니다.");
+    }
+
+    if (!results[1].exists()) {
+      throw new Error("댓글이 존재하지 않습니다.");
+    }
+
     const addReportCountPromise = await updateDoc(commentDoc, {
       reportCount: increment(1),
-      isBlock: reportCommentData.reportCount >= 4 ? true : false
+      isBlock: reportCommentData.reportCount >= 4 ? true : false,
+      reportUidList: arrayUnion(auth.currentUser.uid)
     });
 
     const decreaseCommentCountPromise = [];
@@ -200,17 +300,8 @@ export const fetchReportComment = async (
         })
       );
     }
-    if (!auth.currentUser) return;
-    const userDoc = doc(db, `user/${auth.currentUser.uid}`);
-    const addReportListPromise = updateDoc(userDoc, {
-      reportCommentList: arrayUnion(reportCommentData.commentId)
-    });
 
-    await Promise.all([
-      addReportCountPromise,
-      decreaseCommentCountPromise,
-      addReportListPromise
-    ]);
+    await Promise.all([addReportCountPromise, decreaseCommentCountPromise]);
   } catch (error) {
     console.error(error);
     throw error;
